refactor(homePage): move HomePageCard to dedicated folder
- Moved HomePageCard.jsx to src/components/homePage/ for better file organization.
- Updated import path in HomePage.jsx accordingly.
