+++ src/components/homePage/HomePageCard.jsx
@@ -1,6 +1,7 @@
 import { RecipeImage } from "src/components/RecipeImage";
 
 export const HomePageCard = ({ item }) => {
+
   return (
     <div className="group flex w-[350px] flex-col items-center gap-3 rounded-2xl bg-white300 py-4 text-white ring-2 ring-white ring-offset-2 md:w-[300px] lg:w-[290px]">
       <div className="h-[250px] w-[300px] overflow-hidden rounded-3xl border-[3px] border-orange600 hover:border-orange md:h-[280px] md:w-[280px]">
